+++ sic-storage-web/src/main/java/ru/insoft/archive/sic_storage/model/view/VStrgDocContents.java
@@ -36,13 +36,13 @@ public class VStrgDocContents implements JsonOut
     private String series;
 
     @Column(name = "CASE_COUNT")
-    private Integer caseCount;
+    private Long caseCount;
 
-	public Integer getCaseCount() {
+	public Long getCaseCount() {
 		return caseCount;
 	}
 
-	public void setCaseCount(Integer caseCount) {
+	public void setCaseCount(Long caseCount) {
 		this.caseCount = caseCount;
 	}
